Add roms_ext.h. upload roms
added ch8_ext_roms.h which has a few roms packed in. the idea is that i can remove the file reading code from the examples and just have a quick menu array which allows the user to select a built in rom.

made a quick script called roms2h for packing all the roms.

included a link to the public domain roms.